start rain of arrows tree, reorganization
