+++ Content/SkillPassives/RainOfArrowsTree/NaturesBarrage.cs
@@ -1,0 +1,6 @@
+﻿using PathOfTerraria.Common.Mechanics;
+using PathOfTerraria.Common.Systems.Skills;
+
+namespace PathOfTerraria.Content.SkillPassives.RainOfArrowsTree;
+
+internal class NaturesBarrage(SkillTree tree) : SkillSpecial(tree);
