+++ Content/SkillSpecials/FireNova.cs
@@ -3,6 +3,4 @@ using PathOfTerraria.Common.Systems.Skills;
 
 namespace PathOfTerraria.Content.SkillSpecials;
 
-internal class FireNova(SkillTree tree) : SkillSpecial(tree)
-{
-}
+internal class FireNova(SkillTree tree) : SkillSpecial(tree);
