+++ Content/SkillSpecials/IceNova.cs
@@ -3,6 +3,4 @@ using PathOfTerraria.Common.Systems.Skills;
 
 namespace PathOfTerraria.Content.SkillSpecials;
 
-internal class IceNova(SkillTree tree) : SkillSpecial(tree)
-{
-}
+internal class IceNova(SkillTree tree) : SkillSpecial(tree);
