+++ Content/SkillSpecials/LightningNova.cs
@@ -3,6 +3,4 @@ using PathOfTerraria.Common.Systems.Skills;
 
 namespace PathOfTerraria.Content.SkillSpecials;
 
-internal class LightningNova(SkillTree tree) : SkillSpecial(tree)
-{
-}
+internal class LightningNova(SkillTree tree) : SkillSpecial(tree);
