+++ Content/SkillTrees/NovaTree.cs
@@ -1,5 +1,7 @@
 ﻿using PathOfTerraria.Common.Systems.Skills;
 using PathOfTerraria.Content.SkillPassives;
+using PathOfTerraria.Content.SkillPassives.Generic;
+using PathOfTerraria.Content.SkillPassives.NovaTree;
 using PathOfTerraria.Content.Skills.Magic;
 using PathOfTerraria.Content.SkillSpecials;
