+++ Localization/en-US/Mods.PathOfTerraria.SkillSpecials.hjson
@@ -14,3 +14,8 @@ LightningNova: {
 	Name: Thunder Shock
 	Description: Create a pulse of electricity based on your held magic weapon's damage, dealing a wider range of damage and inflicting Shocked
 }
+
+NaturesBarrage: {
+	Name: Nature's Barrage
+	Tooltip: Tampered tips poison enemies
+}
